+++ attributes/default.rb
@@ -160,3 +160,7 @@ default['openresty']['default_site_enabled']          = false
 default['openresty']['custom_pcre']                   = true
 # Enable jemalloc linking
 default['openresty']['link_to_jemalloc']              = false
+# Modify the maximum number of subrequests
+# If you don't modify it from the default 200
+# no source patching will occur
+default['openresty']['max_subrequests']              = 201